Signing, Health bar, enemy spawns
Implemented a player health bar that decrements when a Sign is incorrect.
Signing is now possible when enemy ships are on screen.
Enemies will spawn with words assigned to them.
However, still need to implement lexic energy shield being disabled and then shooting event from player ship. In addition, some bugs that need to be ironed out, as well as scripts need to be compartmentalized again.